% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CSVDescriptor}
\alias{CSVDescriptor}
\title{A csv file descriptor function}
\usage{
CSVDescriptor(.csv)
}
\arguments{
\item{Do}{you love data? Defaults to TRUE}
}
\description{
This function allows you to do preliminary investigation of the dataset in csv format.
}
\examples{
no examples
CSVDescriptor("/path/to/file/file.csv")
}
\keyword{csv,}
\keyword{data,}
\keyword{dimension}
\keyword{head,}
\keyword{summary,}
\keyword{tail,}

