% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{DescribeCSV}
\alias{DescribeCSV}
\title{A csv file descriptor function}
\usage{
DescribeCSV(.csv)
}
\arguments{
\item{Do}{you love data? Defaults to TRUE}
}
\description{
This function allows you to do preliminary investigation of the dataset in csv format.
}
\examples{
no examples
CSVDescriptor("/path/to/file/file.csv")
}
\keyword{csv,}
\keyword{data,}
\keyword{dimension,}
\keyword{head,}
\keyword{pdf}
\keyword{report,}
\keyword{report.}
\keyword{structure,}
\keyword{summary,}
\keyword{tail,}
\keyword{text}

